requirements.txt
aiohttp~=3.7.4.post0
discord.py~=1.7.3
lxml~=4.8.0
requests~=2.25.1
urllib3~=1.25.8
matplotlib~=3.5.1
pytz~=2022.1
discordhealthcheck~=0.0.8
